Rename BandEntity to Band and implemented connection.
